+++ layer-api/src/main/java/org/layer/annotation/DisableSwaggerSecurity.java
@@ -1,11 +1,11 @@
-package org.layer.common.annotation;
+package org.layer.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.PARAMETER)
+@Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface MemberId {
+public @interface DisableSwaggerSecurity {
 }
