+++ layer-api/src/main/java/org/layer/domain/form/controller/FormApi.java
@@ -1,6 +1,6 @@
 package org.layer.domain.form.controller;
 
-import org.layer.common.annotation.MemberId;
+import org.layer.annotation.MemberId;
 import org.layer.domain.form.controller.dto.request.FormNameUpdateRequest;
 import org.layer.domain.form.controller.dto.request.RecommendFormQueryDto;
 import org.layer.domain.form.controller.dto.request.RecommendFormSetRequest;
