+++ layer-api/src/main/java/org/layer/domain/form/controller/FormController.java
@@ -3,7 +3,7 @@ package org.layer.domain.form.controller;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 
-import org.layer.common.annotation.MemberId;
+import org.layer.annotation.MemberId;
 import org.layer.domain.form.controller.dto.request.FormNameUpdateRequest;
 import org.layer.domain.form.controller.dto.request.RecommendFormQueryDto;
 import org.layer.domain.form.controller.dto.request.RecommendFormSetRequest;
